refactor: GameResult 개선
- 파일 위치 변경
- Lose 케이스 함수 분리
- totalDealerProfit 초기값 상수화
