+++ src/main/kotlin/blackjack/domain/result/GameResult.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.result
 
 import blackjack.domain.person.Dealer
 import blackjack.domain.person.Player
@@ -10,37 +10,36 @@ class GameResult(dealer: Dealer, players: List<Player>) {
     val dealerProfit: Double
 
     init {
-        var totalDealerProfit = 0.0
+        var totalDealerProfit = ZERO_PAYOUT
         playerPayouts =
             players.associateWith { player ->
                 val resultState = ResultState.calculateWin(player, dealer)
-                val payout = calculatePayout(player, resultState)
                 totalDealerProfit += calculateDealerProfit(player, resultState)
-                payout
+                calculatePlayerPayout(player, resultState)
             }
         dealerProfit = totalDealerProfit
     }
 
-    private fun calculatePayout(
+    private fun calculatePlayerPayout(
         player: Player,
         resultState: ResultState,
     ): Double {
-        val payout =
-            when (resultState) {
-                ResultState.WIN -> calculateWinPayout(player)
-                ResultState.LOSE -> player.betAmount.toDouble() * LOSING_PAYOUT_MULTIPLIER
-                ResultState.DRAW -> calculateDrawPayout(player)
-            }
-
-        return payout
+        return when (resultState) {
+            ResultState.WIN -> calculateWinPayout(player)
+            ResultState.LOSE -> calculateLosePayout(player)
+            ResultState.DRAW -> calculateDrawPayout(player)
+        }
     }
 
     private fun calculateWinPayout(player: Player): Double {
         if (player.gameState == PersonState.BLACKJACK) {
             return player.betAmount * BLACKJACK_PAYOUT_MULTIPLIER
         }
+        return player.betAmount * DRAW_PAYOUT_MULTIPLIER
+    }
 
-        return (player.betAmount * DRAW_PAYOUT_MULTIPLIER)
+    private fun calculateLosePayout(player: Player): Double {
+        return player.betAmount.toDouble() * LOSING_PAYOUT_MULTIPLIER
     }
 
     private fun calculateDrawPayout(player: Player): Double {
@@ -54,7 +53,6 @@ class GameResult(dealer: Dealer, players: List<Player>) {
         if (resultState == ResultState.LOSE) {
             return player.betAmount.toDouble()
         }
-
         return ZERO_PAYOUT
     }
