+++ src/main/kotlin/blackjack/view/OutputView.kt
@@ -1,10 +1,10 @@
 package blackjack.view
 
-import blackjack.domain.GameResult
 import blackjack.domain.card.Card
 import blackjack.domain.person.Dealer
 import blackjack.domain.person.Person
 import blackjack.domain.person.Player
+import blackjack.domain.result.GameResult
 
 class OutputView : BlackJackOutputView {
     override fun printMessage(message: String?) {
